run_predict.py dafault bf16

+++ paddlemix/examples/qwen_vl/run_predict.py
@@ -105,7 +105,7 @@ class ModelArguments:
         metadata={"help": "save visual image."},
     )
     dtype: str = field(
-        default="float32",
+        default="bfloat16",
         metadata={"help": "dtype,support float32/float16/bfloat16."},
     )
